factor out more methods
