+++ sbb_newspapers_org_image/unused.py
@@ -49,3 +49,35 @@ def resize_ann(seg_in, input_height, input_width):
     """
     return cv2.resize(seg_in, (input_width, input_height), interpolation=cv2.INTER_NEAREST)
 
+def get_one_hot(seg, input_height, input_width, n_classes):
+    seg = seg[:, :, 0]
+    seg_f = np.zeros((input_height, input_width, n_classes))
+    for j in range(n_classes):
+        seg_f[:, :, j] = (seg == j).astype(int)
+    return seg_f
+
+def color_images(seg, n_classes):
+    ann_u = range(n_classes)
+    if len(np.shape(seg)) == 3:
+        seg = seg[:, :, 0]
+
+    seg_img = np.zeros((np.shape(seg)[0], np.shape(seg)[1], 3)).astype(np.uint8)
+    colors = sns.color_palette("hls", n_classes)
+
+    for c in ann_u:
+        c = int(c)
+        segl = seg == c
+        seg_img[:, :, 0] = segl * c
+        seg_img[:, :, 1] = segl * c
+        seg_img[:, :, 2] = segl * c
+    return seg_img
+
+def cleaning_probs(self, probs: np.ndarray, sigma: float) -> np.ndarray:
+    # Smooth
+    if sigma > 0.0:
+        return cv2.GaussianBlur(probs, (int(3 * sigma) * 2 + 1, int(3 * sigma) * 2 + 1), sigma)
+    elif sigma == 0.0:
+        return cv2.fastNlMeansDenoising((probs * 255).astype(np.uint8), h=20) / 255
+    else:  # Negative sigma, do not do anything
+        return probs
+
